commiting

+++ api/internal/repositories/PostRepository.go
@@ -42,7 +42,7 @@ type DBPostRepository struct {
 // InsertPost creates a new post
 func (r DBPostRepository) InsertPost(ctx context.Context, userId int, content string, facets db.Facets, attributes *db.Attributes) (*models.Post, error) {
 	var post, err = r.querier.InsertPost(ctx, queries.InsertPostParams{
-		UserID:     int32(userId),
+		UserID:     userId,
 		Text:       pgtype.Text{String: content, Valid: true},
 		Facets:     facets,
 		Attributes: attributes,
@@ -57,17 +57,17 @@ func (r DBPostRepository) InsertPost(ctx context.Context, userId int, content st
 
 // DeletePost removes a post by ID
 func (r DBPostRepository) DeletePost(ctx context.Context, postId int) error {
-	return r.querier.DeletePost(ctx, int32(postId))
+	return r.querier.DeletePost(ctx, postId)
 }
 
 // GetPostById retrieves a post by ID
 func (r DBPostRepository) GetPostById(ctx context.Context, postId int) (*models.Post, error) {
-	var dbPost, err = r.querier.GetPostById(ctx, int32(postId))
+	var dbPost, err = r.querier.GetPostById(ctx, postId)
 	if err != nil {
 		return nil, err
 	}
 	return &models.Post{
-		PostID:     int(dbPost.PostID),
+		PostID:     dbPost.PostID,
 		UserID:     dbPost.UserID,
 		Text:       dbPost.Text.String,
 		CreatedAt:  dbPost.CreatedAt.Time.UTC(),
@@ -79,44 +79,44 @@ func (r DBPostRepository) GetPostById(ctx context.Context, postId int) (*models.
 // IsPostLikedByUserId checks if a post is liked by a specific user
 func (r DBPostRepository) IsPostLikedByUserId(ctx context.Context, userId int, postId int) (bool, error) {
 	return r.querier.GetIsPostLikedByUser(ctx, queries.GetIsPostLikedByUserParams{
-		PostID: int32(postId),
-		UserID: int32(userId),
+		PostID: postId,
+		UserID: userId,
 	})
 }
 
 // GetImagesForPost retrieves all images for a specific post
 func (r DBPostRepository) GetImagesForPost(ctx context.Context, postId int) ([]queries.Image, error) {
-	return r.querier.GetImagesByPostId(ctx, int32(postId))
+	return r.querier.GetImagesByPostId(ctx, postId)
 }
 
 // GetAllImagesForUser retrieves all images for a specific user
 func (r DBPostRepository) GetAllImagesForUser(ctx context.Context, userId int) ([]queries.Image, error) {
-	return r.querier.GetAllImagesByUserId(ctx, int32(userId))
+	return r.querier.GetAllImagesByUserId(ctx, userId)
 }
 
 // InsertImage adds a new image to a post
 func (r DBPostRepository) InsertImage(ctx context.Context, postId int, height int, width int, url string, displayOrder int) (queries.Image, error) {
 	return r.querier.InsertImage(ctx, queries.InsertImageParams{
-		PostID:       int32(postId),
-		Height:       int32(height),
-		Width:        int32(width),
+		PostID:       postId,
+		Height:       height,
+		Width:        width,
 		ImageBlobUrl: url,
-		DisplayOrder: int32(displayOrder),
+		DisplayOrder: displayOrder,
 	})
 }
 
 // GetCommentCountForPost returns the number of comments for a post
 func (r DBPostRepository) GetCommentCountForPost(ctx context.Context, postId int) (int, error) {
-	count, err := r.querier.GetCommentCountByPostID(ctx, int32(postId))
+	count, err := r.querier.GetCommentCountByPostID(ctx, postId)
 	return int(count), err
 }
 
 // GetAllPostIds retrieves IDs of all posts with pagination
 func (r DBPostRepository) GetAllPostIds(ctx context.Context, limit int, offset int, currentUserId int) ([]int, error) {
 	postIds32, err := r.querier.GetAllPostIds(ctx, queries.GetAllPostIdsParams{
-		Limit:  int32(limit),
-		Offset: int32(offset),
-		UserID: int32(currentUserId),
+		Limit:  limit,
+		Offset: offset,
+		UserID: currentUserId,
 	})
 	if err != nil {
 		return nil, err
@@ -124,7 +124,7 @@ func (r DBPostRepository) GetAllPostIds(ctx context.Context, limit int, offset i
 
 	postIds := make([]int, len(postIds32))
 	for i, id := range postIds32 {
-		postIds[i] = int(id)
+		postIds[i] = id
 	}
 	return postIds, nil
 }
@@ -132,9 +132,9 @@ func (r DBPostRepository) GetAllPostIds(ctx context.Context, limit int, offset i
 // GetPostIdsForFollowing retrieves post IDs from users a specified user follows
 func (r DBPostRepository) GetPostIdsForFollowing(ctx context.Context, userId int, limit int, offset int) ([]int, error) {
 	postIds32, err := r.querier.GetPostIdsByFollowing(ctx, queries.GetPostIdsByFollowingParams{
-		UserID: int32(userId),
-		Limit:  int32(limit),
-		Offset: int32(offset),
+		UserID: userId,
+		Limit:  limit,
+		Offset: offset,
 	})
 	if err != nil {
 		return nil, err
@@ -142,7 +142,7 @@ func (r DBPostRepository) GetPostIdsForFollowing(ctx context.Context, userId int
 
 	postIds := make([]int, len(postIds32))
 	for i, id := range postIds32 {
-		postIds[i] = int(id)
+		postIds[i] = id
 	}
 	return postIds, nil
 }
@@ -150,9 +150,9 @@ func (r DBPostRepository) GetPostIdsForFollowing(ctx context.Context, userId int
 // GetPostIdsForUser retrieves all post IDs for a specific user
 func (r DBPostRepository) GetPostIdsForUser(ctx context.Context, userId int, limit int, offset int) ([]int, error) {
 	postIds32, err := r.querier.GetPostsIdsByUserId(ctx, queries.GetPostsIdsByUserIdParams{
-		UserID: int32(userId),
-		Limit:  int32(limit),
-		Offset: int32(offset),
+		UserID: userId,
+		Limit:  limit,
+		Offset: offset,
 	})
 	if err != nil {
 		return nil, err
@@ -160,7 +160,7 @@ func (r DBPostRepository) GetPostIdsForUser(ctx context.Context, userId int, lim
 
 	postIds := make([]int, len(postIds32))
 	for i, id := range postIds32 {
-		postIds[i] = int(id)
+		postIds[i] = id
 	}
 	return postIds, nil
 }
@@ -168,22 +168,22 @@ func (r DBPostRepository) GetPostIdsForUser(ctx context.Context, userId int, lim
 // GetPostIdsForMutualFeed retrieves post IDs for mutual feed with relationship metadata
 func (r DBPostRepository) GetPostIdsForMutualFeed(ctx context.Context, userId int, limit int, offset int) ([]queries.GetPostIdsForMutualFeedRow, error) {
 	return r.querier.GetPostIdsForMutualFeed(ctx, queries.GetPostIdsForMutualFeedParams{
-		FollowerID: int32(userId),
-		Limit:      int32(limit),
-		Offset:     int32(offset),
+		FollowerID: userId,
+		Limit:      limit,
+		Offset:     offset,
 	})
 }
 
 // GetPollVotesGrouped retrieves poll votes grouped by option index
 func (r DBPostRepository) GetPollVotesGrouped(ctx context.Context, postId int) ([]queries.GetPollVotesGroupedRow, error) {
-	return r.querier.GetPollVotesGrouped(ctx, int32(postId))
+	return r.querier.GetPollVotesGrouped(ctx, postId)
 }
 
 // GetUserVoteInPoll retrieves the user's vote for a specific poll
 func (r DBPostRepository) GetUserVoteInPoll(ctx context.Context, postId int, userId int) (*int, error) {
 	vote, err := r.querier.GetUserVoteInPoll(ctx, queries.GetUserVoteInPollParams{
-		PostID: int32(postId),
-		UserID: int32(userId),
+		PostID: postId,
+		UserID: userId,
 	})
 	if err != nil {
 		if errors.Is(err, pgx.ErrNoRows) {
@@ -191,16 +191,16 @@ func (r DBPostRepository) GetUserVoteInPoll(ctx context.Context, postId int, use
 		}
 		return nil, err
 	}
-	result := int(vote)
+	result := vote
 	return &result, nil
 }
 
 // InsertVote adds a vote for a poll option
 func (r DBPostRepository) InsertVote(ctx context.Context, postId int, userId int, optionIndex int) error {
 	return r.querier.InsertVote(ctx, queries.InsertVoteParams{
-		PostID:      int32(postId),
-		UserID:      int32(userId),
-		OptionIndex: int32(optionIndex),
+		PostID:      postId,
+		UserID:      userId,
+		OptionIndex: optionIndex,
 	})
 }
 
@@ -212,9 +212,9 @@ func (r DBPostRepository) GetAllPostIdsCursor(ctx context.Context, limit int, be
 	}
 
 	postIds32, err := r.querier.GetAllPostIdsCursor(ctx, queries.GetAllPostIdsCursorParams{
-		Limit:   int32(limit),
+		Limit:   limit,
 		Column2: timestamp,
-		UserID:  int32(currentUserId),
+		UserID:  currentUserId,
 	})
 	if err != nil {
 		return nil, err
@@ -222,7 +222,7 @@ func (r DBPostRepository) GetAllPostIdsCursor(ctx context.Context, limit int, be
 
 	postIds := make([]int, len(postIds32))
 	for i, id := range postIds32 {
-		postIds[i] = int(id)
+		postIds[i] = id
 	}
 	return postIds, nil
 }
@@ -235,8 +235,8 @@ func (r DBPostRepository) GetPostIdsForFollowingCursor(ctx context.Context, user
 	}
 
 	postIds32, err := r.querier.GetPostIdsByFollowingCursor(ctx, queries.GetPostIdsByFollowingCursorParams{
-		UserID:  int32(userId),
-		Limit:   int32(limit),
+		UserID:  userId,
+		Limit:   limit,
 		Column3: timestamp,
 	})
 	if err != nil {
@@ -245,7 +245,7 @@ func (r DBPostRepository) GetPostIdsForFollowingCursor(ctx context.Context, user
 
 	postIds := make([]int, len(postIds32))
 	for i, id := range postIds32 {
-		postIds[i] = int(id)
+		postIds[i] = id
 	}
 	return postIds, nil
 }
@@ -258,8 +258,8 @@ func (r DBPostRepository) GetPostIdsForMutualFeedCursor(ctx context.Context, use
 	}
 
 	return r.querier.GetPostIdsForMutualFeedCursor(ctx, queries.GetPostIdsForMutualFeedCursorParams{
-		FollowerID: int32(userId),
-		Limit:      int32(limit),
+		FollowerID: userId,
+		Limit:      limit,
 		Column3:    timestamp,
 	})
 }
@@ -272,8 +272,8 @@ func (r DBPostRepository) GetPostIdsByUserIdCursor(ctx context.Context, userId i
 	}
 
 	postIds32, err := r.querier.GetPostIdsByUserIdCursor(ctx, queries.GetPostIdsByUserIdCursorParams{
-		UserID:  int32(userId),
-		Limit:   int32(limit),
+		UserID:  userId,
+		Limit:   limit,
 		Column3: timestamp,
 	})
 	if err != nil {
@@ -282,7 +282,7 @@ func (r DBPostRepository) GetPostIdsByUserIdCursor(ctx context.Context, userId i
 
 	postIds := make([]int, len(postIds32))
 	for i, id := range postIds32 {
-		postIds[i] = int(id)
+		postIds[i] = id
 	}
 	return postIds, nil
 }
